[Refactor] StickyHeaderView 분리
